Move interface and world declarations.
This commit moves interface and world declarations under type statement, rather
than top-level statements.

This refactoring will help with my first attempt to register types as part of
resolution, as now type statement contains all possible type declarations.
